+++ src/modules/tools/filterGeneral/tests/unit/utils/compileSnippets.spec.js
@@ -2,6 +2,8 @@ import {expect} from "chai";
 import {
     removeInvalidSnippets,
     convertStringSnippetsIntoObjects,
+    addParent,
+    getFlatArrayOfParentsAndChildren,
     addSnippetIds,
     addSnippetAdjustment,
     addSnippetApi,
@@ -56,6 +58,42 @@ describe("src/modules/tools/filterGeneral/utils/compileSnippets.js", () => {
 
             convertStringSnippetsIntoObjects(snippets);
             expect(snippets).to.deep.equal(expected);
+        });
+    });
+    describe("addParent", () => {
+        it("should add the given parent as parent to all given children, should work recursive", () => {
+            const snippets = [
+                {},
+                {children: [
+                    {}
+                ]}
+            ];
+
+            addParent(snippets, "parent");
+            expect(snippets).to.be.an("array").and.to.have.lengthOf(2);
+            expect(snippets[0]).to.be.an("object").and.to.include.all.keys(["parent"]);
+            expect(snippets[1]).to.be.an("object").and.to.include.all.keys(["parent"]);
+            expect(snippets[1].children).to.be.an("array").and.to.have.lengthOf(1);
+            expect(snippets[1].children[0]).to.be.an("object").and.to.include.all.keys(["parent"]);
+        });
+    });
+    describe("getFlatArrayOfParentsAndChildren", () => {
+        it("should hook all found children after their parents into the given array", () => {
+            const snippets = [
+                    {id: 0, children: [
+                        {id: 1}
+                    ]},
+                    {id: 2}
+                ],
+                result = getFlatArrayOfParentsAndChildren(snippets);
+
+            expect(result).to.be.an("array").and.to.have.lengthOf(3);
+            expect(result[0]).to.be.an("object");
+            expect(result[0].id).to.equal(0);
+            expect(result[1]).to.be.an("object");
+            expect(result[1].id).to.equal(1);
+            expect(result[2]).to.be.an("object");
+            expect(result[2].id).to.equal(2);
         });
     });
     describe("addSnippetIds", () => {